Move genesis generation to its own module
